compound packages: some subpackages are not built
some subpackage names contain digits (for example libX11), but
we test that the tarball name with digits replaced with "X"
contain $package-X. The problem is that the digits have been
replaced in the whole tarball name, and don't match $package.
So translate
$package too. Note that there is no problem if some
packages have names only differing with digits, since
we are sure that the tarball name contain "$package-".
We just want to rule out the case when what comes after
$package- is not a version string.